Base dialect extension class
