+++ lighthouse/dialects/transform/smt_ext/dialect.py
@@ -1,4 +1,4 @@
-from mlir.dialects import ext
+from lighthouse.dialects import DialectExtension
 
 
 def register_and_load(**kwargs):
@@ -7,14 +7,7 @@ def register_and_load(**kwargs):
     TransformSMTExtensionDialect.load(**kwargs)
 
 
-class TransformSMTExtensionDialect(ext.Dialect, name="transform_smt_ext"):
+class TransformSMTExtensionDialect(DialectExtension, name="transform_smt_ext"):
     """A Transform Dialect extension for SMT-related operations."""
 
-    @classmethod
-    def load(cls, *args, **kwargs):
-        # Registers the dialect and its op classes and loads the dialect and ops into the context.
-        super().load(*args, **kwargs)
-
-        for op in cls.operations:
-            if hasattr(op, "attach_interfaces"):
-                op.attach_interfaces()
+    pass
